backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
python-dotenv==1.0.1
sqlalchemy==2.0.27
asyncpg==0.29.0
redis==5.0.1
celery==5.3.6
langchain==0.1.9
openai==1.12.0
google-generativeai==0.3.2
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.9
aiohttp==3.9.3
websockets==12.0
pytest==8.0.0
httpx==0.26.0 